wip: Изменил структуру
